LayoutTests/fast/forms/checkValidity-004.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<form method="get" id="sad_form">
<fieldset name="victim"></fieldset>
<input name="victim" type="text" value="invalid" pattern="something"/>
<button name="victim">lorem ipsum</button>
<select name="victim"></select>
<textarea name="victim"></textarea>
</form>
<div id="console"></div>
<script>
description("This test checks if checkValidity() returns correctly a false (meaning there's an invalid element) result on form element.");

f = document.getElementById("sad_form");
shouldBe("f.checkValidity()", "false");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
