Safari-611.1.10/LayoutTests/fast/forms/checkValidity-004-expected.txt

This test checks if checkValidity() returns correctly a false (meaning there's an invalid element) result on form element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 lorem ipsum
PASS f.checkValidity() is false
PASS successfullyParsed is true

TEST COMPLETE
